Add "retrowars-latest" to manage new game additions gracefully.
If we were to setup "retrowars1" to only accept people with the new client, then they will have a poor experience prior to upgrading. They would need to typically wait 10 seconds or so for Heroku to fire up "retrowars2", which is really just a backup in case all else fails.